Rename files & adjust icons
